python-ui/requirements.txt
# SafeErase Python UI Requirements

# Core GUI Framework
tkinter-tooltip==2.1.0
customtkinter==5.2.0
pillow==10.0.1

# System Integration
psutil==5.9.6
platform-system==1.0.8

# Cryptography and Security
cryptography==41.0.7
pycryptodome==3.19.0
hashlib-compat==1.0.1

# Certificate Generation
reportlab==4.0.7
qrcode[pil]==7.4.2
pypdf==3.17.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
json-schema==4.19.2

# HTTP and API
requests==2.31.0
urllib3==2.0.7
aiohttp==3.9.1

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Logging and Monitoring
loguru==0.7.2
rich==13.7.0
tqdm==4.66.1

# Configuration
pyyaml==6.0.1
configparser==6.0.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Platform Specific (Windows)
pywin32==306; sys_platform == "win32"
wmi==1.5.1; sys_platform == "win32"

# Platform Specific (Linux)
python-dbus==1.3.2; sys_platform == "linux"
pygobject==3.46.0; sys_platform == "linux"

# Optional Dependencies
matplotlib==3.8.2
seaborn==0.13.0
